Fix H1: change nSubsidyHalvingInterval from int to int64_t
The field holds 1,042,600,000 and is used in arithmetic with int64_t
nHeight in validation.cpp. Making the type explicit eliminates the
implicit int→int64_t promotion and prevents any future overflow risk.

Activation height fields (BIP16Height etc.) are left as int since they
must match CBlockIndex::nHeight throughout the codebase.

+++ src/test/validation_tests.cpp
@@ -34,7 +34,7 @@ static void TestBlockSubsidyHalvings(const Consensus::Params& consensusParams)
     BOOST_CHECK_EQUAL(GetBlockSubsidy((int64_t)maxHalvings * consensusParams.nSubsidyHalvingInterval + 1, consensusParams), 0);
 }
 
-static void TestBlockSubsidyHalvings(int nSubsidyHalvingInterval)
+static void TestBlockSubsidyHalvings(int64_t nSubsidyHalvingInterval)
 {
     Consensus::Params consensusParams;
     consensusParams.nSubsidyHalvingInterval = nSubsidyHalvingInterval;
